Fixed Cinder test
I think this is not correct that there should be only 1 backend.
The bankends are set in the cinder.conf in "enabled_backends" and
are comma-seperated, so there can be more than 1 backend. E.g. my
customer env has 2 backends.
Fixed the test: there should be >= 1 backend, id est more than 1
backend. "wc -l" always returns the digit (0 or more), so there is
enough to convert the str to int and check >=1.
Tested on the customer's env.