Add timeout parametr for contrail-status
Without this -t parameter, on environments with more api workers it might result that contrail-api timeouted, because default timeout is 2 seconds and with more api workers it is not just enought. It can be bypassed by curling api or just adding parameter above " -t 10 ". Then results will be more accurate.