Added test to check all packages are latest
Added test to check that all packages on the nodes are latest and
are not upgradable. Added the possibility to skip some packages
in global_config if they should not be upgraded.

Problem description:
The 'test_check_package_versions' checks that the versions are
consistent across the nodes of the same group. But we have no
test that the actual versions are really correct and from correct
MCP release. I had the cloud that had some packages installed
from the wrong repositories, not from the required MCP release, so
the fix was to upgrade packages. So there is the need to have
the test to check that the packages are latest.
At the same time if several packages should not be upgraded and
have correct version even if Installed!=Candidate, there is
possibility to skip packages.

Currently the test is skipped by default ("skip_test": True in
global_config.yaml file). Set False to run the test.

Change-Id: Iddfab8b3d7fb4e72870aa0791e9da95a66f0ccfd