Added test to check minions status
The CVP Sanity tests skip the nodes automatically if this minion
does not respond in 1 sec (salt_timeout in config) time.
Sometimes all the tests can pass, but some KVM nodes along with its
Control plane VMs can be down and CVP tests will not test this and
will not inform about this.
The tests check that all minions are up.

Change-Id: Ib8495aeb043448b36aea85bb31ee2650d655075e