[SPT] Added downloaded iperf package for vm2vm test
Added possibility to execute vm2vm test in case if
the OpenStack instances have no Internet to install
iperf packages from Internet using apt.
In case if there is no Internet, the deb package is
already dwnloaded and present inside the Docker image.
When VM is ready, the iperf package is transported to
it and installed.
The possibility is configured with 'internet_at_vms'
string option that is 'true' by default.

Change-Id: I45c47f9691f484340e7f4a9decc584dcedb63b6c
Related-PROD: #PROD-32023