[Sanity] Exclude bmt* nodes from checking nova hypervisors consistency
The test test_nova_hosts_consistent failed on Ironic envs because
bmt* nodes are Ironic client nodes with ironic-* and nova-compute
services running. Due to nova-compute services up, the bmt* nodes
are expected to appear in openstack hosts and nova service list,
however they are not expected to by in hypervisors list.
So, need to skip bmt* nodes in this test.

Change-Id: I76ce24bba7f3fc74c009f6741aee09b6f246bc23
Related-prod: PROD-34170