Fix logic of test_cinder_services_has_all_backends
Currently this test checks if number of the cinder volumes is equal with
number of the backends in pillar. But in fact, some backends could use
more than one link with cinder volume service (or use several
availability zones).

If we check not just equality between numbers, but links between all
available in cluster backends with cinder-volume, it should be more
usefull.

Related-Bug: PROD-36238
Related-Bug: PROD-36737
Change-Id: I53b44e777adbdd0385533cd0fc49bdb7439b26cc